requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
decorator==4.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
instagrapi==2.2.1
moviepy==1.0.3
numpy==2.3.5
pillow==12.0.0
proglog==0.1.12
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.4
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
